Create `/1/_status/surge-predictions/` endpoint
Involved factoring out some code and tests from the status index.

[#87607978]
