+++ api/apps/status/views/__init__.py
@@ -1,1 +1,2 @@
 from .status_index import StatusIndex
+from .surge_predictions import SurgePredictions
