+++ api/apps/status/views/common.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+Status = namedtuple('Status', 'ok,text')
