+++ api/apps/status/views/status_index.py
@@ -7,11 +7,12 @@ from django.shortcuts import render
 from collections import namedtuple
 
 from api.apps.locations.models import Location
-from api.apps.predictions.models import TidePrediction, SurgePrediction
+from api.apps.predictions.models import TidePrediction
 from api.apps.observations.models import Observation
 
 from ..alert_manager import AlertType, is_alert_enabled
-
+from .surge_predictions import check_surge_predictions
+from .common import Status
 
 LocationStatus = namedtuple('LocationStatus', 'name,checks')
 Check = namedtuple('Check', 'name,status_text,status_class')
@@ -53,19 +54,19 @@ def get_location_status(location):
     all_ok = True
     checks = []
     for name, function in checks_to_run:
-        ok, text = function(location)
-        all_ok = all_ok and ok
+        status = function(location)
+        all_ok = all_ok and status.ok
         checks.append(Check(
             name=name,
-            status_text=text,
-            status_class='success' if ok else 'danger'))
+            status_text=status.text,
+            status_class='success' if status.ok else 'danger'))
 
     return all_ok, LocationStatus(name=location.name, checks=checks)
 
 
 def check_tide_predictions(location):
     if not is_alert_enabled(location, AlertType.tide_predictions):
-        return (True, 'OK (alert disabled)')
+        return Status(True, 'OK (alert disabled)')
 
     one_month_away = (datetime.datetime.now(pytz.UTC)
                       + datetime.timedelta(days=30))
@@ -75,37 +76,14 @@ def check_tide_predictions(location):
         minute__datetime__gte=one_month_away).exists()
 
     if ok:
-        return (True, 'OK')
+        return Status(True, 'OK')
     else:
-        return (False, '< 30 days left')
-
-
-def check_surge_predictions(location):
-    """
-    Test that we have a surge prediction for every minute in the next 36 hours.
-    """
-    if not is_alert_enabled(location, AlertType.surge_predictions):
-        return (True, 'OK (alert disabled)')
-
-    now = datetime.datetime.now(pytz.UTC)
-    thirty_seven_hours_away = now + datetime.timedelta(hours=36)
-
-    count = SurgePrediction.objects.filter(
-        location=location,
-        minute__datetime__gte=now,
-        minute__datetime__lt=thirty_seven_hours_away).count()
-    ok = (36 * 60) == count
-
-    if ok:
-        return (True, 'OK')
-    else:
-        return (False,
-                'Missing data for next 36 hours: {} vs 2160'.format(count))
+        return Status(False, '< 30 days left')
 
 
 def check_observations(location):
     if not is_alert_enabled(location, AlertType.observations):
-        return (True, 'OK (alert disabled)')
+        return Status(True, 'OK (alert disabled)')
 
     one_hour_ago = (datetime.datetime.now(pytz.UTC)
                     - datetime.timedelta(minutes=120))
@@ -115,6 +93,6 @@ def check_observations(location):
         minute__datetime__gte=one_hour_ago).exists()
 
     if ok:
-        return (True, 'OK')
+        return Status(True, 'OK')
     else:
-        return (False, '> 2 hours old')
+        return Status(False, '> 2 hours old')
